Always hide protected apps from the recent tasks list
If a protected app is started by a protected component manager, it
will be available in the recent tasks list. This allows to bypass
the protection since restarting an app from recents doesn't require
authentication. Fix this changing the flags of the activity so that
it's never added to the recent tasks list.

This change also ensures that apps don't add entries to the recent
tasks list: a protected component that starts a new instance of
itself it's treated like a protected component manager and therefore
its entries are not removed from the recent tasks list.

Change-Id: Ib44b392cc9c27ec2b7c8f16735c38ee45011be3e