Fix to android 7.0+ Intent Flag Bug
On Android 7.0+ versions intent flag is broken.

FLAG_ACTIVITY_NEW_TASK is no longer required as
a result of an incorrect if statement in startActivity.
When startActivity(Intent intent) is called it will
internally call startActivity(Intent intent, Bundle options)
while null as the second argument.

Since Android 7.0+ has changed how the conditional checking is
done in the second method, an intent that does not have
FLAG_ACTIVITY_NEW_TASK added to the Intent and calls
startActivity through the first method will NEVER hit a
runTimeException as it should and does in pre 7.0 Android
since we will short-circuit out of the if statement when
we evaluate && options != null.

A new suggestion is to check the flag first, then nest the
conditionals in order to allow both cases to be accounted for.

This issue is tracked at https://issuetracker.google.com/issues/64069975

Change-Id: Ia08dcef042c2d5d18bd781fb8999c4b2d1422fdb
Signed-off-by: zsmoore <zsmoore@zachary-moore.com>
Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>